+++ lib/rubocop/cop/style/if_unless_modifier.rb
@@ -4,8 +4,8 @@ module RuboCop
   module Cop
     module Style
       # Checks for if and unless statements that would fit on one line
-      # if written as a modifier if/unless.
-      # The maximum line length is configurable.
+      # if written as a modifier if/unless. The maximum line length is
+      # configured in the `Metrics/LineLength` cop.
       class IfUnlessModifier < Cop
         include StatementModifier