+++ spec/rubocop/cop/style/while_until_modifier_spec.rb
@@ -110,22 +110,4 @@ describe RuboCop::Cop::Style::WhileUntilModifier do
       RUBY
     end
   end
-
-  context 'when the maximum line length is specified by the cop itself' do
-    let(:config) do
-      hash = {
-        'Metrics/LineLength' => { 'Max' => 100 },
-        'Style/WhileUntilModifier' => { 'MaxLineLength' => 80 }
-      }
-      RuboCop::Config.new(hash)
-    end
-
-    it "accepts multiline while that doesn't fit on one line" do
-      check_too_long(cop, 'while')
-    end
-
-    it "accepts multiline until that doesn't fit on one line" do
-      check_too_long(cop, 'until')
-    end
-  end
 end